Add WinForms project.
Add WinForms project with MainForm and SelectProcessForm. MainForm
displays the SelectProcessForm immediately for the user to select a
process. If the user does not select a process and has not selected a
process in the past the MainForm prompts to exit - if the user selects
yes to exit the application closes, otherwise it presents the
SelectProcessForm again.

This logic will allow the user to exit when they desire or if they
press the wrong keyboard shortcut on the SelectProcessForm will prevent
the annoyance of having to restart the application. The Dialogue may
be an annoyance so an option to never ask me again may be preferred -
but could lead to an endless loop.

An event should be created in the SelectProcessForm that the MainForm
can register to when a Process is selected, this will allow the
MainForm to be notified when the user has selected a Process instead of
listening to the DialogResult. I am not sure which is better practice.
In the case of the event, additional logic may be required to ensure
the original thread receives a notification so that an exception is
not thrown about a second thread attempting to update the UI.