Move WinForms to Forms subdirectory. Create SelectProcessFormViewModel.
Move MainForm and SelectProcessForm to Forms keep root of project
clean. Update references to use new location for Forms. This will keep
the project clean and organized.

Add DataGridView to display process information to user. Add Process Id
to SelectProcessFormViewModel - which is displayed in DataGridView.
Doing so will allow the user to see something at this time.

Additional fields will need to be added to SelectProcessFormViewModel
and additional logic will be required to determine that selected
process is valid before allowing user to select the Process. Buttons to
refresh the view and to accept/cancel the form will need to be provided.
An event when the process is selected is required. Additional form
configuration is missing.