Rename SelectProcessForm Files, Display Process Info
Rename SelectProcessForm to ProcessSelectorForm to give it a purpose
instead of a noun. Customized the DataGridViewProcess. Add and
customized buttons to Refresh, Select, and Cancel - added shortcut keys
that can be displayed with Alt.

Add Name and WindowTitle to the ProcessSelectorFormViewModel and tied
the BindingList to the DataGridViewProcess.DataSource so the data gets
displayed.

Still need to add the event when a Process is selected. Additional
logic will be required to determine if a Process is valid. Need to
decide where that logic belongs. Should the MainForm determine if a
process is valid and show it to the user, or should the
ProcessSelectorForm determine if a process is valid before it closes.
Additionally, what should happen if the Process closes after the user
has selected it and gone back to the MainForm?