Add Events for Closing, CellDoubleClick, and ColumnHeaderMouseClick.
Add FormClosing event to check if the user cancelled. This will allow
the form to check if the Process is still valid before sending it to
the MainForm.

Add CellDoubleClick to count as selecting a process. This is an ease
of use feature for the user. Instead of selecting a process, mousing
over to the Select button, we allow a double click that will simulate
the click for them. Custom logic had to be included to not detect the
cell resizing double click.

Add ColumnHeaderMouseClick which will allow the user to sort by the
column header. Additional code will be needed to sort the BindingList
as well as considering the filter requirements that also should be
included.