Add ProcessSelect event and add logic to FormExiting to validate Process before selecting
Add ProcessSelect event to ProcessSelectorForm. Allows the user to
notify the MainForm (or other) when a Process has been selected by the
user.

Add validation to ensure process is valid before giving Process to the
MainForm (or other). Notifies the user if the Process has been killed
and is no longer a valid selection.