Add Process Info to MainForm
Add labels to show Id, Name, Nonpaged System Memory, Paged System
Memory, Paged Memory, Peak Paged Memory, Virtual Memory, Peak Virtual
Memory to give some information about the selected process to the user.

The memory will need to be updated. A timer could be set and every tick
would refresh the selected process and the memory values. Whether or
not the tick amount should be configurable remains - perhaps this can
be a value in the App.config.

MainForm is still missing PrivateMemorySize and the MainWindowTitle
with the ability to change the value of MainWindowTitle - this should
also be considered being put in the Timer's tick process. It will give
a way to refresh the MainWindowTitle value in case something else has
altered the title.

Still need a way to pop up the ProcessSelectorForm if the Process
closes while in the MainForm. Perhaps periodic checks to
Process.HasExited when the Timer ticks? Or we could subscribe to the
Process.Exited event if we set the Process.EnableRaisingEvents to true.