Add WindowTitle and PrivateMemory
Add WindowTitle to values shown for a Process. Add button that will
allow user to update the WindowTitle.

Add PrivateMemorySize to the values shown for a Process.

Still needs the timer used to refresh the process values. Still needs
the binding set up to allow the user to update the Window Title.