Add Doc comments and MessageBox constants when a process is not selected
Add Docstyle comments describing what a method or variable is for.
Still need docstyle comments for the events, but these should be
copied from the original event.

Add text message contents to the MessageBox that is shown when a user
does not select a process. Made the values private constant members to
prevent recreation and to make the parameters small enough to be
readable in the application.